backend/prediction/requirements.txt
tensorflow-macos==2.15.0
pandas==2.1.4
numpy==1.24.3
scikit-learn==1.3.2
flask==3.0.0
joblib==1.3.2 